requirements.txt
adafruit_circuitpython_dht==3.7.8
board==1.0
gTTS==2.3.1
paho_mqtt==1.6.1
playsound==1.3.0
smbus==1.1.post2
w1thermsensor==2.0.0
paho-mqtt~=1.6.1
pywebview~=4.0.2
redis~=4.5.1
python-dotenv~=1.0.0
qrcode~=7.4.2